move ingest/parse-pdr code into parse-pdr task
